Add controller/view for Krikri::Activity; add PrettifierHelperBehavior
This was a self-serving thing I built out to make it easier for me to find
and inspect instances of `Krikri::Activity` as I wrote the SPARQL queries for QA.

This also moves two methods from `Krikri::SearchResultsHelperBehavior` into a new
module, `Krikri::PrettifierHelperBehavior`.
