+++ app/helpers/krikri/activities_helper.rb
@@ -1,0 +1,5 @@
+module Krikri
+  module ActivitiesHelper
+    include Krikri::PrettifierHelperBehavior
+  end
+end
