+++ app/views/krikri/activities/index.html.erb
@@ -1,0 +1,52 @@
+<% content_for :title do %>Activities<% end %>
+
+<div id="main-container" class="container">
+  <%= render :partial=>'/flash_msg', layout: 'shared/flash_messages' %>
+  <h1>Activities</h1>
+  <%= paginate @activities, theme: 'twitter-bootstrap-3' %>
+
+  <div class="panel panel-default">
+    <table class="table table-bordered table-striped table-hover">
+      <thead>
+        <tr>
+          <th>
+            ID
+          </th>
+          <th>
+            Agent
+          </th>
+          <th>
+            Start Time
+          </th>
+          <th>
+            End Time
+          </th>
+          <th>
+            Options
+          </th>
+        </tr>
+      </thead>
+      <tbody>
+        <% @activities.each do |activity| %>
+          <tr data-item-id=<%= "#{activity.id}" %> class="item">
+            <td>
+              <%= activity.id %>
+            </td>
+            <td>
+              <%= activity.agent %>
+            </td>
+            <td>
+              <%= activity.start_time %>
+            </td>
+            <td>
+              <%= activity.end_time %>
+            </td>
+            <td>
+               <pre class="pre-scrollable"><%= prettify_json_string(activity.opts) %></pre>
+            </td>
+          </tr>
+        <% end %>
+      </tbody>
+    </table>
+  </div>
+</div>
